Add loading of Game and Gameset objects from both files and the database.
Reworks how the DatabaseAccess, Gameset, and Game classes are coupled.

Add a stub of send_gameset_update in the DatabaseAccess object, as well as annotating functions in GameSet to automatically call send_gameset_update when a change is made. This will later be connected to a ToolsManager object that will update the views of the tools that will be reading those gamesets.