Modify how the panel layout works, allowing for much bette control fo the layout.
NOTE: There is still one smaller known issue. Sometimes, if a container goes away via exiting out tabs, it will not clear the space that had been occupied and doesn't seem to actually close the Container, leading to minor visual bugs and not having a visual way to open a new tab.